[Fix] back double click & Serializable image data class

+++ network/src/main/java/com/awilab/network/model/MovieDetail.kt
@@ -170,7 +170,7 @@ data class MovieDetail(
             @SerialName("iso_639_1")
             val iso6391: String? = "",
             @SerialName("vote_average")
-            val voteAverage: Int? = 0,
+            val voteAverage: Double? = 0.0,
             @SerialName("vote_count")
             val voteCount: Int? = 0,
             @SerialName("width")
@@ -190,7 +190,7 @@ data class MovieDetail(
             @SerialName("iso_639_1")
             val iso6391: String? = "",
             @SerialName("vote_average")
-            val voteAverage: Int? = 0,
+            val voteAverage: Double? = 0.0,
             @SerialName("vote_count")
             val voteCount: Int? = 0,
             @SerialName("width")
